Implement `character_delimited` framing
