--- lib/codec/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "codec"
-version = "0.1.0"
-authors = ["Vector Contributors <vector@timber.io>"]
-edition = "2018"
-publish = false
-license = "MPL-2.0"
-
-[dependencies]
-bytes = "1.1.0"
-tokio-util = { version = "0.6.7", features = ["codec"] }
-tracing = "0.1.26"
-
-[dev-dependencies]
-serde_json = "1.0.67"
